requirements.txt
boto3
click
streamlit
python-dotenv